pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rfdetr-tooling"
version = "0.1.0"
description = "RF-DETR finetuning tooling for COCO/YOLO datasets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rfdetr>=1.4.0",
    "torch>=2.7.0,<=2.8.0",
    "torchvision>=0.22.0",
    "pandas>=2.2.0",
    "pydantic>=2.10.0",
    "loguru>=0.7.3",
    "supervision>=0.25.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "Pillow>=10.0.0",
    "cveta2",
    "tensorboard>=2.18.0",
]

[project.scripts]
rfdetr-tool = "rfdetr_tooling.cli:main"

[project.optional-dependencies]
clearml = ["clearml>=1.16.0"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.0.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # missing-trailing-comma - conflicts with ruff formatter
    "ISC001",  # single-line-implicit-string-concatenation - conflicts with formatter
    "EM",      # flake8-errmsg - overly pedantic for small projects
    "TRY003",  # raise-vanilla-args - same as EM
    "D213",    # multi-line-summary-second-line - conflicts with D212
    "D203",    # one-blank-line-before-class - conflicts with D211
    "PLR2004", # magic-value-comparison
    "RUF002", # ambiguous-unicode-character-docstring - Russian comments
    "RUF003", # ambiguous-unicode-character-comment - Russian comments
]
fixable = ["ALL"]

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["rfdetr.*", "supervision.*", "pycocotools.*", "clearml.*"]
ignore_missing_imports = true

[tool.uv.sources]
cveta2 = { path = "../cveta2", editable = true }

